disease-identifier-app/backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
opencv-python-headless
pillow
python-multipart
tf-explain